Move Value and Variable
